upload download changed added search for uploaded files

+++ app/controllers/upload_controller.rb
@@ -2,18 +2,9 @@ class UploadController < ApplicationController
   before_action :logged_in_user, only: [:upload_download_page]
   before_action :user_library, only: [:uploadFile, :deleteDOC, :deletePDF]
   def upload_download_page
-    @PDFfiles=Hash.new
-    @DOCfiles=Hash.new
-    @files = Dir.glob('public/digital_books/*')
-    for f in @files
-      id = f.to_s.from(21).to_i
-      file = DataFile.find id
-      if(file.name.include? '.pdf')
-        @PDFfiles[id] = file.name
-      elsif(file.name.include? '.doc')
-         @DOCfiles[id] = file.name
-      end
-    end
+
+    @files = DataFile.search(params[:search])
+
   end
 
   def index
@@ -48,21 +39,34 @@ class UploadController < ApplicationController
   end
 
   def deletePDF
+
     name = params[:filename]
-    directory = "public/digital_books"
-    File.delete(Rails.root.join(directory, name+'.pdf')) if File.file? Rails.root.join(directory, name+'.pdf')
-    message = "File deleted successfully"
-    flash[:success] = message
-    redirect_to upload_upload_download_page_path
-  end
+    f = DataFile.find name
+    if f.destroy
+      directory = "public/digital_books"
+      File.delete(Rails.root.join(directory, name+'.pdf')) if File.file? Rails.root.join(directory, name+'.pdf')
+      message = "File deleted successfully"
+      flash[:success] = message
+      redirect_to upload_upload_download_page_path
+    else
+      message = "Eror while deleting file"
+      flash[:danger] = message
+    end
+end
 
   def deleteDOC
     name = params[:filename]
-    directory = "public/digital_books"
-    File.delete(Rails.root.join(directory, name+'.doc'))
-    message = "File deleted successfully"
-    flash[:success] = message
-    redirect_to upload_upload_download_page_path
+    f = DataFile.find name
+    if f.destroy
+      directory = "public/digital_books"
+      File.delete(Rails.root.join(directory, name+'.doc'))
+      message = "File deleted successfully"
+      flash[:success] = message
+      redirect_to upload_upload_download_page_path
+    else
+      message = "Eror while deleting file"
+      flash[:danger] = message
+    end
   end
